// src/types.ts
export type Filter = {
    name: string,
    value: string
}